Changed the way the zoom level is handled in the code.  Previously, we had a hard-coded array of doubles that set the pixel scale (called pixelScale[]), and zooming in and out would [dec/inc]rement the index of the array.  We now simply store a double for the current pixelScale (called ZoomFactor), and zooming in/out changes this value by multiplying/dividing by a preset factor DZOOM.
At this point, DZOOM is #defined in kstarsdata.h to be 1.1; it could
conceivably be an Advanced view option.  The smaller DZOOM is, the smoother
zooming will be (but there will also be more steps).

Doing Zoom this way will allow us to set the ZoomFactor arbitrarily, so that
a "Zoom-to-box" function can be implemented.  It would also be nice to have
a DCOP function "setZoom( double )".

svn path=/trunk/kdeedu/kstars/; revision=239275

+++ kstars/starobject.h
@@ -73,7 +73,7 @@ class StarObject : public SkyObject {
   */
 		StarObject( double r, double d, double m, QString n="unnamed",
   					 		QString n2="", QString sptype="--" );
-          
+
 /**
 	*Constructor.  Set SkyObject data according to arguments.
 	*@param t Type of object
@@ -121,11 +121,11 @@ class StarObject : public SkyObject {
   *Returns empty string if star has no genetive name defined.
   *@returns greek letter portion of genetive name
   */
-		QString greekLetter( void );		
-		
+		QString greekLetter( void );
+
 		//overloaded from SkyObject
-		void drawLabel( QPainter &psky, int x, int y, int zoom, bool drawName, bool drawMag, double scale );
-		
+		void drawLabel( QPainter &psky, int x, int y, double zoom, bool drawName, bool drawMag, double scale );
+
 	private:
 		QString SpType;
 		SkyObjectName *soName;
